% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_data.R
\docType{data}
\name{block_census_data}
\alias{block_census_data}
\title{Block-level Decennial Census counts}
\format{A tibble with 39148 rows and 11 variables:
\describe{
\item{geoid}{Census Bureau GEOID}
\item{pop_total}{Total population; P001001}
\item{med_age}{Median age; P013001}
\item{pop_white, pop_white_pct}{Non-hispanic white population; P005003}
\item{pop_black, pop_black_pct}{Non-hispanic black population; P005004}
\item{pop_asian, pop_asian_pct}{Non-hispanic asian population; P005006}
\item{pop_hisp, pop_hisp_pct}{Hispanic, any race population; P004003}
}}
\source{
\url{https://factfinder.census.gov/faces/nav/jsf/pages/index.xhtml}
}
\usage{
block_census_data
}
\description{
A dataset containing U.S. Census Bureau Decennial Census population and race
counts for all census blocks in New York City. Counts are from the 2010
Decennial Census. Margins of error are not included because this data
represents an enumeration of the entire population, rather than a sample.
}
\keyword{datasets}
