% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\docType{data}
\name{block_sf}
\alias{block_sf}
\title{NYC census block boundaries}
\format{An sf object with 38796 rows and 14 variables:
\describe{
\item{geoid}{Census Bureau GEOID; can be used to join spatial data with
Census estimates}
\item{boro_tract_block_id}{NYC DCP borough code and census tract number}
\item{state_fips}{ANSI state FIPS code}
\item{county_fips}{ANSI county FIPS code}
\item{block_id}{Census Bureau block number}
\item{tract_id}{Census Bureau tract number}
\item{county_name}{County name}
\item{boro_name}{Borough name}
\item{boro_id}{NYC DCP borough code and census tract number}
\item{nta_id}{NYC neighborhood tabulation area id}
\item{nta_name}{NYC neighborhood tabulation area name}
\item{puma_id}{Census Bureau public use microdata area id}
\item{puma_name}{Census Bureau public use microdata area name}
\item{geometry}{sfc_MULTIPOLYGON \cr
NAD83 / New York Long Island (ftUS); EPSG:2263}
}}
\source{
\url{https://www1.nyc.gov/site/planning/data-maps/open-data/districts-download-metadata.page}
}
\usage{
block_sf
}
\description{
A simple features (sf) dataset containing the geometry of all census blocks
in New York City from the 2010 Census.
}
\keyword{datasets}
