% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\docType{data}
\name{borough_sf}
\alias{borough_sf}
\alias{borough_sf_simple}
\title{NYC borough boundaries}
\format{An sf object with 5 rows and 8 variables:
\describe{
\item{geoid}{Census Bureau GEOID; can be used to join spatial data with
Census estimates}
\item{state_fips}{ANSI state FIPS code}
\item{county_fips}{ANSI county FIPS code}
\item{county_name}{County name}
\item{boro_name}{Borough name}
\item{boro_id}{NYC DCP borough code and census tract number}
\item{geometry}{sfc_MULTIPOLYGON \cr
NAD83 / New York Long Island (ftUS); EPSG:2263}
}}
\source{
\url{https://www1.nyc.gov/site/planning/data-maps/open-data/districts-download-metadata.page}
}
\usage{
borough_sf

borough_sf_simple
}
\description{
A simple features (sf) dataset containing the geometry of the five boroughs
of New York City. The boroughs of New York City are equivalent to counties.
}
\keyword{datasets}
