% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\docType{data}
\name{cd_sf}
\alias{cd_sf}
\alias{cd_sf_simple}
\title{NYC community district boundaries}
\format{An sf object with 71 rows and 9 variables:
\describe{
\item{boro_cd_id}{NYC borough and community district id; unique ids}
\item{cd_id}{NYC community district id; ids repeat across boroughs}
\item{cd_name}{NYC community district names}
\item{state_fips}{ANSI state FIPS code}
\item{county_fips}{ANSI county FIPS code}
\item{county_name}{County name}
\item{boro_name}{Borough name}
\item{boro_id}{NYC DCP borough code and census tract number}
\item{geometry}{sfc_MULTIPOLYGON \cr
NAD83 / New York Long Island (ftUS); EPSG:2263}
}}
\source{
\url{https://www1.nyc.gov/site/planning/data-maps/open-data.page#other}
}
\usage{
cd_sf

cd_sf_simple
}
\description{
A simple features (sf) dataset containing the geometry of all community
districts (CDs) and joint interest areas in New York City.
}
\keyword{datasets}
