% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\docType{data}
\name{cong_sf}
\alias{cong_sf}
\alias{cong_sf_simple}
\title{U.S Congressional district boundaries in NYC}
\format{An sf object with 13 rows and 5 variables:
\describe{
\item{geoid}{Census Bureau GEOID; can be used to join spatial data with
Census estimates}
\item{council_dist_id}{NYC city council district id}
\item{council_dist_name}{NYC city council district name}
\item{state_fips}{ANSI state FIPS code}
\item{geometry}{sfc_MULTIPOLYGON \cr
NAD83 / New York Long Island (ftUS); EPSG:2263}
}}
\source{
\url{https://www1.nyc.gov/site/planning/data-maps/open-data.page#other}
}
\usage{
cong_sf

cong_sf_simple
}
\description{
A simple features (sf) dataset containing the geometry of all U.S.
Congressional districts in New York City.
}
\keyword{datasets}
