% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_blocks.R
\name{nyc_blocks}
\alias{nyc_blocks}
\title{NYC census block boundaries}
\usage{
nyc_blocks(filter_by = NULL, region = NULL, add_census_data = FALSE)
}
\arguments{
\item{filter_by}{The geography to filter by. Possible values are \code{borough, nta, puma}. If \code{NULL}, all census blocks are returned.}

\item{region}{A character vector of boroughs, NTAs, or PUMAs. Selected
regions much match the geography indicated by \code{filter_by} argument.}

\item{add_census_data}{If \code{TRUE}, selected demographic, social, and economic
data from the U.S. Census 2010 decennial census is appended to block
boundaries.}

\item{resolution}{The resolution of the map. Defaults to lower resolution.}
}
\value{
An \code{sf} object of census block boundaries
}
\description{
Get current (2010) boundaries for census blocks in New York City in the
simple features {sf} format. Either get all census blocks or subset of blocks
filtered by borough, neighborhood tabulation area (NTA), or public use
microdata area (PUMA).
}
\details{
For more information about the metadata included with boundaries,
see \link{blocks_sf}.
}
\examples{
if (require(sf)) {

  # get sf boundaires
  all_nyc_blocks <- nyc_blocks()

  north_si_blocks <- nyc_blocks(
    filter_by = "nta",
    region = c("SI22", "SI35"),
    resolution = "high"
    )

  queens_blocks <- nyc_blocks(
    filter_by = "borough",
    region = "Queens",
    add_acs_data = TRUE
    )

  # plot boundaries
  plot(st_geometry(all_nyc_blocks))

  plot(st_geometry(queens_blocks))

  plot(queens_blocks["med_hhinc_est"])
}

}
