% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_boros.R
\name{nyc_boros}
\alias{nyc_boros}
\title{NYC borough (county) boundaries}
\usage{
nyc_boros(region = NULL, add_acs_data = FALSE, resolution = c("low",
  "high"))
}
\arguments{
\item{region}{A character vector of boroughs. If \code{NULL}, all boroughs are
returned.}

\item{add_acs_data}{If \code{TRUE}, selected demographic, social, and economic
data from the U.S. Census Bureau American Community Survey is appended to
borough boundaries.}

\item{resolution}{The resolution of the map. Defaults to lower resolution.}
}
\value{
An \code{sf} object of borough boundaries
}
\description{
Get current boundaries for public use boroughs (counties) in New York City in
the simple features {sf} format. Either get all boroughs or a subset of
boroughs.
}
\details{
For more information about the metadata included with boundaries,
see \link{boros_sf}.
}
\examples{
if (require(sf)) {

  # get sf boundaires
  all_nyc_boros <- nyc_boros()

  mn_bx_boros <- nyc_boros(
    region = c("mahattan", "bronx"),
    resolution = "high"
    )

  nyc_boros_acs <- nyc_boros(add_acs_data = TRUE)

  # plot boundaries
  plot(st_geometry(all_nyc_boros))

  plot(st_geometry(mn_bx_boros))

  plot(nyc_boros_acs["pop_hisp_est"])
}
}
