% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_cds.R
\name{nyc_cds}
\alias{nyc_cds}
\title{NYC community district boundaries}
\usage{
nyc_cds(filter_by = NULL, region = NULL, resolution = c("low",
  "high"))
}
\arguments{
\item{filter_by}{The geography to filter by. Possible values are
"borough" or "cd". If \code{NULL}, all CDs are returned.}

\item{region}{A character vector of boroughs or CDs. Selected regions much
match the geography indicated by \code{filter_by} argument.}

\item{resolution}{The resolution of the map. Defaults to lower resolution.}
}
\value{
An \code{sf} object of CD boundaries.
}
\description{
Get current boundaries for community districts (CDs) and joint interest areas
in New York City in the simple features {sf} format. Either get all CDs or a
subset of CDs filtered by borough.
}
\details{
For more information about the metadata included with boundaries,
see \link{cds_sf}.
}
\examples{
if (require(sf)) {

  # get sf boundaires
  all_nyc_cds <- nyc_cds()

  queens_brooklyn_cds <- nyc_cds(
    filter_by = "boro",
    region = c("queens", "brooklyn")
  )

  # plot boundaries
  plot(st_geometry(all_nyc_cds))

  plot(st_geometry(queens_brooklyn_cds))
}
}
