% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_ntas.R
\name{nyc_ntas}
\alias{nyc_ntas}
\title{NYC neighborhood tabulation area boundaries}
\usage{
nyc_ntas(filter_by = NULL, region = NULL, add_acs_data = FALSE,
  resolution = c("low", "high"))
}
\arguments{
\item{filter_by}{The geography to filter by. Possible values are "borough",
"puma", or "nta". If \code{NULL}, all NTAs are returned.}

\item{region}{A character vector of boroughs, PUMAs, or NTAs. Selected
regions much match the geography indicated by \code{filter_by} argument.}

\item{add_acs_data}{If \code{TRUE}, selected demographic, social, and economic
data from the U.S. Census Bureau American Community Survey is appended to
NTA boundaries.}

\item{resolution}{The resolution of the map. Defaults to lower resolution.}
}
\value{
An \code{sf} object of NTA boundaries
}
\description{
Get current boundaries for neighborhood tabulation areas (NTAs) in New York
City in the simple features {sf} format. Either get all NTAs or a subset of
NTAs filtered by borough or public use microdata area (PUMA).
}
\details{
For more information about the data fields included with boundaries,
see \link{ntas_sf}.
}
\examples{
if (require(sf)) {

  # get sf boundaires
  all_nyc_ntas <- nyc_ntas()

  greenpoint_williamsburg_nta <- nyc_ntas(
    filter_by = "puma",
    region = "4001",
    resolution = "high"
    )

  queens_brooklyn_ntas <- nyc_ntas(
    filter_by = "boro",
    region = c("queens", "brooklyn"),
    add_acs_data = TRUE
  )

  # plot boundaries
  plot(st_geometry(all_nyc_ntas))

  plot(st_geometry(greenpoint_williamsburg_nta))

  plot(queens_brooklyn_ntas["pop_white_pct_est"])
}
}
