% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-in-poly.R
\name{nyc_point_poly}
\alias{nyc_point_poly}
\title{Find administrative districts for a set of points}
\usage{
nyc_point_poly(points, geography)
}
\arguments{
\item{points}{An \code{sf} object containing only point features.}

\item{geography}{A character vector of geographies to. Possible values are
"borough", "puma", "nta", "tract", or "block".}
}
\value{
An \code{sf} object of point features with administrative district codes appended.
}
\description{
blah blah blah
}
\details{
blag blag
}
\examples{
if (require(sf)) {
  points <- st_sf(geometry = st_sample(nyc_boros(), 100))
  nyc_point_poly(points, c("boro", "nta", "block"))
}

}
