% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_pumas.R
\name{nyc_pumas}
\alias{nyc_pumas}
\title{NYC public use microdata area boundaries}
\usage{
nyc_pumas(filter_by = NULL, region = NULL, add_acs_data = FALSE,
  resolution = c("low", "high"))
}
\arguments{
\item{filter_by}{The geography to filter by. Possible values are "borough" or
"puma". If \code{NULL}, all PUMAs are returned.}

\item{region}{A character vector of boroughs or PUMAs. Selected regions much
match the geography indicated by \code{filter_by} argument.}

\item{add_acs_data}{If \code{TRUE}, selected demographic, social, and economic
data from the U.S. Census Bureau American Community Survey is appended to
PUMA boundaries.}

\item{resolution}{The resolution of the map. Defaults to lower resolution.}
}
\value{
An \code{sf} object of PUMA boundaries
}
\description{
Get current boundaries for public use microdata area (PUMAs) in New York City
in the simple features {sf} format. Either get all PUMAs or a subset of PUMAs
filtered by borough or PUMA.
}
\details{
For more information about the metadata included with boundaries,
see \link{pumas_sf}.
}
\examples{
if (require(sf)) {

  # get sf boundaires
  all_nyc_pumas <- nyc_pumas()

  greenpoint_williamsburg_puma <- nyc_pumas(
    filter_by = "puma",
    region = "4001",
    resolution = "high"
    )

  queens_brooklyn_pumas <- nyc_pumas(
    filter_by = "boro",
    region = c("queens", "brooklyn"),
    add_acs_data = TRUE
  )

  # plot boundaries
  plot(st_geometry(all_nyc_pumas))

  plot(st_geometry(greenpoint_williamsburg_puma))

  plot(queens_brooklyn_pumas["med_age_est"])
}
}
