% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_tract.R, R/tracts.R
\docType{data}
\name{nyc_tract}
\alias{nyc_tract}
\alias{tracts_simple}
\title{NYC census tract boundaries}
\format{\itemize{
\item 
}}
\usage{
nyc_tract(filter_by = NULL, region = NULL, add_acs_data = FALSE,
  resolution = c("low", "high"))

tracts_simple
}
\arguments{
\item{filter_by}{The geography to filter by. Possible values are \code{borough, nta, puma}. If \code{NULL}, all census tracts are returned.}

\item{region}{A character vector of boroughs, NTAs, or PUMAs. Selected
regions much match the geography indicated by \code{filter_by} argument.}

\item{add_acs_data}{If \code{TRUE}, selected demographic, social, and economic
data from the U.S. Census Bureau American Community Survey is appended to
tract boundaries.}

\item{resolution}{The resolution of the map. Defaults to lower resolution.}
}
\value{
An \code{sf} object of census tract boundaries
}
\description{
Get current (2010) boundaries for census tracts in New York City in the
simple features {sf} format. Either get all census tracts or subset of tracts
filtered by borough, neighborhood tabulation area (NTA), or public use
microdata area (PUMA).
}
\details{
For more information about the data fields included with boundaries,
see \link{tracts}.
}
\examples{
if (require(sf)) {

  # get sf boundaires
  all_nyc_tracts <- nyc_tract()

  north_si_tracts <- nyc_tract(
    filter_by = "nta",
    region = c("SI22", "SI35"),
    resolution = "high"
    )

  queens_tracts <- nyc_tract(
    filter_by = "borough",
    region = "Queens",
    add_acs_data = TRUE
    )

  # plot boundaries
  plot(st_geometry(all_nyc_tracts))

  plot(st_geometry(queens_tracts))

  plot(queens_tracts["med_hhinc"])
}

}
\keyword{datasets}
