% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_tracts.R
\name{nyc_tracts}
\alias{nyc_tracts}
\title{NYC census tract boundaries}
\usage{
nyc_tracts(filter_by = NULL, region = NULL, resolution = c("low",
  "high"))
}
\arguments{
\item{filter_by}{The geography to filter by. Possible values are \code{borough, nta, puma}. If \code{NULL}, all census tracts are returned.}

\item{region}{A character vector of boroughs, NTAs, or PUMAs. Selected
regions much match the geography indicated by \code{filter_by} argument.}

\item{resolution}{The resolution of the map. Defaults to lower resolution.}
}
\value{
An \code{sf} object of census tract boundaries
}
\description{
Get current (2010) boundaries for census tracts in New York City in the
simple features {sf} format. Either get all census tracts or subset of tracts
filtered by borough, neighborhood tabulation area (NTA), or public use
microdata area (PUMA).
}
\examples{
if (require(sf)) {

  # get sf boundaires
  all_nyc_tracts <- nyc_tracts()
  queens_tracts <- nyc_tracts(filter_by = "borough", region = "Queens")
  north_si_tracts <- nyc_tracts(
    filter_by = "nta",
    region = c("SI22", "SI35"),
    resolution = "high"
    )

  # plot boundaries
  plot(st_geometry(all_nyc_tracts))
  plot(st_geometry(queens_tracts))
  plot(st_geometry(queens_tracts))
}

}
