% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycgeo-package.R
\docType{package}
\name{nycgeo-package}
\alias{nycgeo}
\alias{nycgeo-package}
\title{nycgeo: Spatial Data Files for NYC}
\description{
Spatial data files for various geographic
    boundaries in New York City.  Data is in the in the sf (simple
    features) format and includes boundaries for boroughs (counties),
    neighborhood tabulation areas, community districts, public microdata
    areas, census tracts, and census blocks.
}
\author{
\strong{Maintainer}: Matt Herman \email{mfherman@gmail.com}

}
\keyword{internal}
