% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\docType{data}
\name{police_sf}
\alias{police_sf}
\alias{police_sf_simple}
\title{NYC police precinct boundaries}
\format{An sf object with 77 rows and 4 variables:
\describe{
\item{police_precinct_id}{NYC police precinct id}
\item{police_precinct_name}{NYC police precinct name}
\item{state_fips}{ANSI state FIPS code}
\item{geometry}{sfc_MULTIPOLYGON \cr
NAD83 / New York Long Island (ftUS); EPSG:2263}
}}
\source{
\url{https://www1.nyc.gov/site/planning/data-maps/open-data.page#other}
}
\usage{
police_sf

police_sf_simple
}
\description{
A simple features (sf) dataset containing the geometry of all police
precincts in New York City.
}
\keyword{datasets}
