% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\docType{data}
\name{school_sf}
\alias{school_sf}
\alias{school_sf_simple}
\title{NYC school district boundaries}
\format{An sf object with 32 rows and 4 variables:
\describe{
\item{school_dist_id}{NYC school district id}
\item{school_dist_name}{NYC school district name}
\item{state_fips}{ANSI state FIPS code}
\item{geometry}{sfc_MULTIPOLYGON \cr
NAD83 / New York Long Island (ftUS); EPSG:2263}
}}
\source{
\url{https://www1.nyc.gov/site/planning/data-maps/open-data.page#other}
}
\usage{
school_sf

school_sf_simple
}
\description{
A simple features (sf) dataset containing the geometry of all community
school districts in New York City.
}
\keyword{datasets}
