% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts_acs_data.R
\docType{data}
\name{tracts_acs_data}
\alias{tracts_acs_data}
\title{Tract-level American Community Survey estimates}
\format{A tibble with 2177 rows and 30 variables:
\describe{
\item{geoid}{Census Bureau GEOID}
\item{pop_total}{Total population; B01001_001}
\item{med_age}{Median age; B01002_001}
\item{med_hhinc}{Median annual household income (dollars); B19013_001}
\item{pop_white, pop_white_pct}{Non-hispanic white population; B03002_003}
\item{pop_black, pop_black_pct}{Non-hispanic black population; B03002_004}
\item{pop_hisp, pop_hisp_pct}{Hispanic, any race population; B03002_012}
\item{pop_asian, pop_asian_pct}{Non-hispanic asian population; B03002_006}
\item{pop_ba_above, pop_ba_above_pct}{Population 25 years or older with at
least a Bachelor's degree; B15003_022, B15003_023, B15003_024, B15003_025}
\item{pop_inpov, pop_inpov_pct}{Population with income below poverty line;
B17021002}
}}
\source{
\url{https://factfinder.census.gov/faces/nav/jsf/pages/index.xhtml}
}
\usage{
tracts_acs_data
}
\description{
A dataset containing U.S. Census Bureau American Community Survey (ACS)
estimates and margins of error of selected demographic, social, and economic
variables for all census tracts in New York City. Variables are five-year
estimates from the 2012-2017 ACS.
}
\keyword{datasets}
