% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_point_poly.R
\name{nyc_point_poly}
\alias{nyc_point_poly}
\title{Find which administrative districts a set of points are within}
\usage{
nyc_point_poly(points, geography)
}
\arguments{
\item{points}{An \code{sf} object containing only point features.}

\item{geography}{A character vector of administrative boundaries to append to
points. Possible values are "borough", "puma", "cd" "nta", "tract", and
"block". If more than one geography is selected, names and codes for each
geography is appended.}
}
\value{
An \code{sf} object of point features with administrative district names
and and codes appended.
}
\description{
Point-in-polygon operation for NYC administrative boundaries. Use
\code{nyc_point_poly()} on a set of points to determine which administrative
district each point lies within.
}
\details{
Using \code{nyc_point_poly()} requires installing the \href{"https://r-spatial.github.io/sf/"}{sf package}.
}
\examples{
if (require(sf)) {

  # generate 100 random points in nyc
  points <- st_sf(geometry = st_sample(nyc_boundaries(), 100))

  nyc_point_poly(points = points, geography = c("borough", "nta", "block"))
}

}
