% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycgeo-package.R
\docType{package}
\name{nycgeo-package}
\alias{nycgeo}
\alias{nycgeo-package}
\title{nycgeo: Spatial Data Files for NYC}
\description{
Spatial data files for various geographic and administrative
    boundaries in New York City as well as tools for working with NYC spatial
    data. Data is in the in the sf (simple features) format and includes
    boundaries for boroughs (counties), public use microdata areas, community
    districts, neighborhood tabulation areas, census tracts, and census blocks. 
    In addition, selected demographic, social, and economic estimates from the 
    U.S. Census Bureau American Community Survey are included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mfherman/nycgeo}
  \item Report bugs at \url{https://github.com/mfherman/nycgeo/issues}
}

}
\author{
\strong{Maintainer}: Matt Herman \email{mfherman@gmail.com}

}
\keyword{internal}
